homework-lesson-10/requirements.txt
# ── lesson-8 runtime dependencies ──────────────────────────────────────────
chromadb>=0.5.0
pypdf>=5.0.0
langchain-text-splitters>=1.0.0
ddgs>=7.0
langchain>=1.2.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.4.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
trafilatura>=2.0.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# ── test / evaluation extras ────────────────────────────────────────────────
deepeval>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
